feat: Add 2 new statistics calculators, move a few things around
[#OCD-3663]

+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/statistics/CriterionListingCountStatisticEntity.java
@@ -31,7 +31,7 @@ public class CriterionListingCountStatisticEntity {
 
     @Basic(optional = false)
     @Column(name = "listing_count", nullable = false)
-    private Integer listingCount;
+    private Long listingCount;
 
     @Basic(optional = false)
     @Column(name = "certification_criterion_id", nullable = false)
